Add simple test for rust diamond
